Compress images with pngquant
The png export functionality provided by vips isn't good enough for images that are hard to compress, failing to reach the 512 KB minimum file size even when compression is set to 10.

If the default export fails to reach the file limit at  compressionLevel=6, we switch over to pngquant and export with that instead.